requirements.txt
botogram2
pillow
